requirements.txt
Flask==1.1.1
uWSGI==2.0.18
Werkzeug==0.16.0
Flask-Cors==3.0.8
requests==2.22.0
pandas==0.25.3
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
opencv-python-headless==4.1.2.30
Pillow==7.0.0
python-dotenv==0.10.5
scikit-image==0.16.2
twilio==6.35.2
ibm-watson==4.2.1